+++ src/Component/Utils/Split/CustomSplit/style.css.ts
@@ -14,7 +14,7 @@ export const contentWrapperStyle = style({
   justifyContent: 'flex-start',
   gap: '1.5rem',
   padding: '2rem 0 2rem 2rem',
-  minWidth: '20rem',
+  minWidth: '15rem',
 });
 
 export const contentWrapperSideStyle = styleVariants({
